requirements.txt
fastapi
uvicorn
paramiko
beautifulsoup4
requests
python-dotenv